requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
alembic==1.12.1
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openai==1.3.5
langchain==0.0.335
pinecone-client==2.2.4
pytesseract==0.3.10
pillow==10.1.0
pandas==2.1.3
numpy==1.25.2
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
celery==5.3.4
flower==2.0.1
python-telegram-bot==20.7
smtplib-ssl==1.0.4
pydantic-settings==2.0.3
asyncpg==0.29.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1